Make cocotb.types implementation modules private
