+++ docs/source/newsfragments/4623.change.rst
@@ -1,0 +1,1 @@
+Renamed ``cocotb.types.ArrayLike`` to :class:`.AbstractArray`.
