+++ src/cocotb/types/_array.py
@@ -3,13 +3,13 @@
 # SPDX-License-Identifier: BSD-3-Clause
 from typing import Iterable, Iterator, List, TypeVar, Union, cast, overload
 
-from cocotb.types import ArrayLike
-from cocotb.types.range import Range
+from cocotb.types._abstract_array import AbstractArray
+from cocotb.types._range import Range
 
 T = TypeVar("T")
 
 
-class Array(ArrayLike[T]):
+class Array(AbstractArray[T]):
     r"""Fixed-size, arbitrarily-indexed, homogeneous collection type.
 
     Arrays are similar to, but different from Python :class:`list`\ s.
