+++ src/cocotb/types/_logic.py
